pyproject.toml
[tool.poetry]
name = "resti"
version = "0.1.0"
description = "Simplified CRM on Django"
authors = ["Denis <ad.way@mail.ru>"]
readme = "README.md"
packages = [{include = "_project_"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-bootstrap5 = "^24.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
